Update integration tests to use any-charm instead of hello-kubecon (#14)
The [any-charm](https://github.com/canonical/any-charm) charm allows us to define any type of relation we wish to test. It is also dedicated to that and has no real world use.  
This makes it a better test target than [hello-kubecon](github.com/jnsgruk/hello-kubecon) for the content-cache-k8s-operator.

This replaces hello-kubecon in the integration tests by a minimal use of any-charm.